pyproject.toml
#:schema https://json.schemastore.org/pyproject.json
[project]
name = "thermompnn-fp"
version = "0.1.0"
description = "FP-style ThermoMPNN reproduction for protein stability prediction"
readme = "README.md"
authors = [
    { name = "Nazelto.WSL", email = "1500465005@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "pandas>=2.2.3",
    "biopython>=1.84",
    "pyyaml>=6.0.2",
    "torch",
    "lightning>=2.5.0",
    "torchmetrics>=1.6.0",
    "tqdm>=4.67.1",
]

[project.scripts]
thermompnn-fp = "thermompnn_fp.cli:main"

[build-system]
requires = ["uv_build>=0.9.29,<0.10.0"]
build-backend = "uv_build"
